Add copyright/patent issues to possible NACK reasons
Adding in response to a Slack discussion where someone was unclear on the fact
that a NACK may be justified if code can't be accepted due to copyright/patent
issues.  For example, it would be reasonable and prudent to NACK a contribution
of AGPL-licensed consensus code on the basis that the license terms are
incompatible with the MIT license used by the rest of the codebase.